refactor: abstract to own components
